+++ src/yasbd/utils/pysbd_adapter.py
@@ -24,11 +24,8 @@ class TextSpan:
         return f"[{self.start}:{self.end}] {self.sent}"
 
     def __eq__(self, other) -> bool:
-        if (
-            isinstance(other, TextSpan)
-            or hasattr(other, "sent")
-            and hasattr(other, "start")
-            and hasattr(other, "end")
+        if isinstance(other, TextSpan) or (
+            hasattr(other, "sent") and hasattr(other, "start") and hasattr(other, "end")
         ):
             return (self.start, self.end, self.sent) == (
                 other.start,
@@ -36,6 +33,9 @@ class TextSpan:
                 other.sent,
             )
         return NotImplemented
+
+    def __hash__(self) -> int:
+        return hash((self.sent, self.start, self.end))
 
 
 class Segmenter:
@@ -124,11 +124,9 @@ class Segmenter:
             return res
 
         if self.clean:
-            sents = list(self._detector.segment(text))
-            return sents
-        else:
-            sents = list(self._detector.segment(text, preserve_whitespace=True))
-            return self._convert_leading_space_to_trails(sents)
+            return list(self._detector.segment(text))
+        sents = list(self._detector.segment(text, preserve_whitespace=True))
+        return self._convert_leading_space_to_trails(sents)
 
     @validate_input
     def sentences_with_char_spans(self, sentences: list[str]) -> list[TextSpan]: